pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-modal>=0.0.1",
    "langchain-google-genai==4.2.0",
    "langgraph-cli[inmem]>=0.4.12",
    "mcp[sse]>=1.26.0",
    "tavily>=1.1.0",
]
